requirements.txt
aniso8601==9.0.1
bcrypt==4.0.0
boto3==1.24.70
botocore==1.27.70
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
